refactor: simplify top-level deploy workflow
